requirements.txt
dnspython==2.4.2
flask==3.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
rich==13.7.0
